api-server/requirements.txt
hug==2.2.0
redis==2.10.5
requests==2.12.4
pyyaml
fakeredis
locustio==0.8a2
mock